Move favorite to separated component
