+++ Apps/hoppy-hub/src/app/beers/beer-details/beer-details.component.html
@@ -66,21 +66,7 @@
           <strong>Composition:</strong> {{ beer.composition }}
         </p>
 
-        @if (!favoriteLoading) {
-          <button
-            type="button"
-            class="btn btn-outline btn-warning flex h-auto items-center justify-self-end p-4"
-            (click)="onToggleFavorite()">
-            <fa-icon class="text-yellow-400" [icon]="faStarr"></fa-icon>
-            <span>
-              @if (!favorite) {
-                Add To Favorites
-              } @else {
-                Remove From Favorites
-              }
-            </span>
-          </button>
-        }
+        <app-favorite [beer]="beer" [user]="user"></app-favorite>
       </div>
 
       <app-beer-opinions [beer]="beer" [user]="user"></app-beer-opinions>
