+++ Apps/hoppy-hub/src/app/beers/beer-details/favorite/favorite.component.html
@@ -1,0 +1,15 @@
+@if (!loading) {
+  <button
+    type="button"
+    class="btn btn-outline btn-warning flex h-auto items-center justify-self-end p-4"
+    (click)="onToggleFavorite()">
+    <fa-icon class="text-yellow-400" [icon]="faStarr"></fa-icon>
+    <span>
+      @if (!favorite) {
+        Add To Favorites
+      } @else {
+        Remove From Favorites
+      }
+    </span>
+  </button>
+}
